skill-swap-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy>=2.0.23
pydantic[email]>=2.5.0
python-multipart==0.0.6
tabulate==0.9.0
gunicorn==21.2.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
